requirements.txt
python-decouple==3.6
termcolor==2.0.1
